Create IWindowBackend and GLFW implementation
Since the beginning of time, Etterna has used a custom window implementation on Windows, macOS, and Linux. This PR will focus on rewriting all code related to the graphics backend, as well as making it easily extendable.

This commit adds an implementation for using GLFW to only initialized a OpenGL context. IWindowBackend would be used as the base class should any other windowing backend be desirable.